pyproject.toml
[tool.poetry]
name = "py-fitbit"
version = "0.1.0"
description = "Python bindings for Fitbit Web API"
authors = ["Team <j.rossi@uva.nl>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/j-rossi-nl/py-fitbit"
repository = "https://github.com/j-rossi-nl/py-fitbit"
documentation = "https://github.com/j-rossi-nl/py-fitbit"
classifiers = [
    "Development Status :: 3 - Alpha",
]
packages = [
    { include="fitbit", from="src"}
]

[tool.poetry.urls]
Changelog = "https://github.com/j-rossi-nl/py-fitbit/releases"

[tool.poetry.dependencies]
python = "^3.8"
click = ">=8.0.1"
certifi = ">=14.05.14"
six = ">=1.10"
python-dateutil = ">=2.5.3"
setuptools = ">=21.0.0"
urllib3 = ">=1.15.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
Pygments = ">=2.10.0"
black = ">=21.10b0"
coverage = ">=4.0.3"
darglint = ">=1.8.1"
flake8 = ">=4.0.1"
flake8-bandit = ">=2.1.2"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.10.1"
mypy = ">=0.930"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pytest = ">=6.2.5"
pyupgrade = ">=2.29.1"
safety = ">=1.10.3"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
typeguard = ">=2.13.3"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
myst-parser = {version = ">=0.16.1"}
nose = ">=1.3.7"
pluggy = ">=0.3.1"
py = ">=1.4.31"
randomize = ">=0.13"
ipykernel = "^6.9.2"
CherryPy = "^18.6.1"
nox-poetry = "^0.9.0"

[tool.poetry.scripts]
py-fitbit = "fitbit.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["fitbit", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 0

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
